feat: explore page - basic
