+++ apps/web/src/app/(landing)/explore/page.tsx
@@ -1,0 +1,7 @@
+import { Explore } from "../../../components/landing/explore/explore";
+
+const ExplorePage = () => {
+  return <Explore />;
+};
+
+export default ExplorePage;
