+++ apps/web/src/components/landing/explore/explore.tsx
@@ -1,0 +1,9 @@
+export const Explore = () => {
+  return (
+    <div className="flex w-full flex-col items-center justify-start gap-14 pb-16 lg:gap-24 lg:pb-28">
+      <h1 className="lg:leading-tighter max-w-4xl text-center text-4xl font-bold tracking-tighter md:text-5xl xl:text-6xl">
+        Explore possibilities
+      </h1>
+    </div>
+  );
+};
