+++ apps/web/src/config/index.ts
@@ -25,6 +25,7 @@ export const REPOSITORY_URL = "https://github.com/Bartek532/syncreads";
 export const TWITTER_URL = "https://twitter.com/bzagrodzki";
 
 export const LANDING_HEADER_NAVIGATION = [
+  { name: "Explore", href: "/explore" },
   { name: "Pricing", href: "/pricing" },
   { name: "FAQ", href: "/#faq" },
   { name: "Contact", href: "/#contact" },
@@ -81,8 +82,8 @@ export const LANDING_OPEN_SOURCE = {
   title: "Create the future together!",
   description: `We believe in transparency and the power of the community, so if you have any ideas or suggestions, feel free to give us a signal. We're open to any feedback!`,
   cta: {
-    href: REPOSITORY_URL,
-    text: "See on Github",
+    href: "/#contact",
+    text: "Give feedback",
   },
 };
 
@@ -121,96 +122,6 @@ export const LANDING_CONTACT = {
   title: "Contact",
   description: `If you have any questions or suggestions, please don't hesitate to drop us a message.`,
 };
-
-export const PRICING_PLANS = [
-  {
-    name: "Starter",
-    price: {
-      month: 0,
-      year: 0,
-    },
-    popular: true,
-    trial: false,
-    description: "Start consuming now. No credit card required",
-    features: [
-      {
-        available: true,
-        new: false,
-        title: "Seamless sync",
-      },
-      {
-        available: true,
-        new: false,
-        title: "reMarkable support",
-      },
-      {
-        available: true,
-        new: true,
-        title: "Kindle support",
-      },
-      {
-        available: true,
-        new: false,
-        title: "Multiple formats (pdf, epub)",
-      },
-      {
-        available: true,
-        new: false,
-        title: "Feeds management",
-      },
-      {
-        available: true,
-        new: false,
-        title: "Realtime tracking (logs, history)",
-      },
-      {
-        available: true,
-        new: false,
-        title: "Simple analytics",
-      },
-    ],
-    cta: {
-      title: "Get started",
-      link: "/auth/register",
-    },
-  },
-  // {
-  //   name: "Pro",
-  //   price: {
-  //     month: 15,
-  //     year: 109,
-  //   },
-  //   popular: true,
-  //   trial: true,
-  //   description: "For small teams and personal projects",
-  //   features: [
-  //     "Everything in Starter",
-  //     "Custom Domain",
-  //     "Advanced Functions",
-  //     "Advanced Database",
-  //     "Advanced Analytics",
-  //     "Priority Support",
-  //   ],
-  // },
-  // {
-  //   name: "Enterprise",
-  //   price: {
-  //     month: 29,
-  //     year: 309,
-  //   },
-  //   popular: false,
-  //   trial: false,
-  //   description: "For large teams and businesses",
-  //   features: [
-  //     "Everything in Pro",
-  //     "Enterprise Functions",
-  //     "Enterprise Database",
-  //     "Enterprise Analytics",
-  //     "SLA",
-  //     "Dedicated Support",
-  //   ],
-  // },
-];
 
 export const TOASTER_CONFIG: DefaultToastOptions = {
   className: "!bg-background !text-primary",
